some reformatting, more functions to come
